backend/requirements.txt
Flask-RESTful==0.3.8
PyJWT==1.7.1
Flask-SQLAlchemy==3.0.3
python-dotenv